// ui/src/api/generated/models/model-parameter-definition-info.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ModelParameterDefinitionInfo
 */
export interface ModelParameterDefinitionInfo {
    /**
     * 
     * @type {boolean}
     * @memberof ModelParameterDefinitionInfo
     */
    'common'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ModelParameterDefinitionInfo
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelParameterDefinitionInfo
     */
    'displayName'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelParameterDefinitionInfo
     */
    'domain'?: ModelParameterDefinitionInfoDomainEnum;
    /**
     * 
     * @type {string}
     * @memberof ModelParameterDefinitionInfo
     */
    'field'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelParameterDefinitionInfo
     */
    'modelType'?: ModelParameterDefinitionInfoModelTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ModelParameterDefinitionInfo
     */
    'parameter'?: string;
}

export const ModelParameterDefinitionInfoDomainEnum = {
    Language: 'language',
    Embedding: 'embedding',
    Rerank: 'rerank',
    ImageGeneration: 'imageGeneration'
} as const;

export type ModelParameterDefinitionInfoDomainEnum = typeof ModelParameterDefinitionInfoDomainEnum[keyof typeof ModelParameterDefinitionInfoDomainEnum];
export const ModelParameterDefinitionInfoModelTypeEnum = {
    Language: 'language',
    Embedding: 'embedding',
    Rerank: 'rerank',
    ImageGeneration: 'image-generation'
} as const;

export type ModelParameterDefinitionInfoModelTypeEnum = typeof ModelParameterDefinitionInfoModelTypeEnum[keyof typeof ModelParameterDefinitionInfoModelTypeEnum];